Development

+++ server/src/controllers/productController.js
@@ -196,29 +196,9 @@ const createProduct = async (req, res) => {
       }
     });
 
-    if (
-      !name ||
-      !description ||
-      !cat_id ||
-      !content ||
-      !brand ||
-      !status ||
-      !weight ||
-      !height ||
-      !price ||
-      !productType ||
-      !slug ||
-      !statusStock
-    ) {
-      uploadModel.deleteImg(thumbnail);
-      uploadModel.deleteImgs(imagesProduct);
-      uploadModel.deleteImgs(imageVariantsC);
-      return res.status(StatusCodes.BAD_REQUEST).json({
-        message: ERROR_MESSAGES.REQUIRED,
-      });
-    }
     const newTags = JSON.parse(tags);
     const newProductType = JSON.parse(productType);
+
     const parsedVars = JSON.parse(variants);
     imageVariantsC.forEach((file, index) => {
       parsedVars[index].image = file;
@@ -246,20 +226,17 @@ const createProduct = async (req, res) => {
     const dataProduct = await productModel.createProduct(data);
 
     if (dataProduct.error) {
-      await uploadModel.deleteImg(thumbnail);
-      await uploadModel.deleteImgs(imagesProduct);
-      await uploadModel.deleteImgs(imageVariantsC);
+      uploadModel.deleteImg(thumbnail);
+      uploadModel.deleteImgs(imagesProduct);
+      uploadModel.deleteImgs(imageVariantsC);
       return res.status(StatusCodes.BAD_REQUEST).json(dataProduct.detail);
     }
     return res.status(StatusCodes.OK).json({ dataProduct });
   } catch (error) {
     if (req.files) {
-      const thumbnail = path.join(
-        'uploads/products',
-        req.files['thumbnail'][0].filename
-      );
+      const thumbnailPath = req.files['thumbnail']?.[0]?.filename;
 
-      let imagesProduct = req.files['images'].map((file) => {
+      const imageVariantsC = req.files['imageVariants']?.map((file) => {
         if (file && file.filename) {
           return path.join('uploads/products', file.filename);
         } else {
@@ -267,20 +244,35 @@ const createProduct = async (req, res) => {
         }
       });
 
-      let imageVariantsC = req.files['imageVariants'].map((file) => {
+      const imagesProduct = req.files['images']?.map((file) => {
         if (file && file.filename) {
           return path.join('uploads/products', file.filename);
         } else {
           throw new Error('File image không hợp lệ');
         }
       });
-      await uploadModel.deleteImg(thumbnail);
-      await uploadModel.deleteImgs(imagesProduct);
-      await uploadModel.deleteImgs(imageVariantsC);
+
+      if (thumbnailPath) {
+        const thumbnail = path.join('uploads/products', thumbnailPath);
+        uploadModel.deleteImg(thumbnail);
+      }
+
+      if (imagesProduct && imagesProduct.length > 0) {
+        uploadModel.deleteImgs(imagesProduct);
+      }
+
+      if (imageVariantsC && imageVariantsC.length > 0) {
+        uploadModel.deleteImgs(imageVariantsC);
+      }
     }
-    return res
-      .status(StatusCodes.INTERNAL_SERVER_ERROR)
-      .json({ message: error.message });
+    if (error.details) {
+      return res.status(StatusCodes.BAD_REQUEST).json({
+        message: error.details[0].message,
+      });
+    }
+    return res.status(StatusCodes.INTERNAL_SERVER_ERROR).json({
+      message: 'Có lỗi xảy ra xin thử lại sau',
+    });
   }
 };
 
@@ -336,43 +328,13 @@ const updateProduct = async (req, res) => {
         }
       });
     }
-
-    if (
-      !name ||
-      !description ||
-      !cat_id ||
-      !content ||
-      !brand ||
-      !status ||
-      !weight ||
-      !height ||
-      !statusStock ||
-      !slug ||
-      !variants
-    ) {
-      if (thumbnail) {
-        await uploadModel.deleteImg(product.thumbnail);
-      }
-      if (imagesProduct && imagesProduct.length > 0) {
-        await uploadModel.deleteImgs(imagesProduct);
-      }
-      if (imageVariantsC && imageVariantsC.length > 0) {
-        await uploadModel.deleteImgs(imageVariantsC);
-      }
-      return res.status(StatusCodes.BAD_REQUEST).json({
-        message: ERROR_MESSAGES.REQUIRED,
-      });
-    }
-
     const product = await productModel.getProductById(id);
 
     let parsedVariants = [];
     let oldImageVariants = [];
-
     if (Array.isArray(variants)) {
       parsedVariants = variants.map((variant, index) => {
         let parsedVariant = JSON.parse(variant);
-
         if ('imageAdd' in parsedVariant) {
           if ('image' in parsedVariant) {
             if (product.variants[index].image === parsedVariant.image) {
@@ -381,7 +343,6 @@ const updateProduct = async (req, res) => {
           }
           delete parsedVariant.imageAdd;
         }
-
         return parsedVariant;
       });
     } else {
@@ -443,14 +404,13 @@ const updateProduct = async (req, res) => {
 
     const newimgURLs = [...validImgs, ...imagesProduct];
     const newThumbnail = thumbnail ? thumbnail : product.thumbnail;
-
     const data = {
       cat_id,
       name: name,
       slug,
       description,
       content,
-      tags,
+      tags: JSON.parse(tags),
       thumbnail: newThumbnail,
       images: newimgURLs,
       brand,
@@ -459,59 +419,36 @@ const updateProduct = async (req, res) => {
       weight,
       height,
       statusStock,
-      productType,
+      productType: JSON.parse(productType),
     };
-    const dataProduct = await productModel.update(id, data);
 
-    if (dataProduct.error) {
-      if (thumbnail) {
-        await uploadModel.deleteImg(product.thumbnail);
-      }
-      if (imagesProduct && imagesProduct.length > 0) {
-        await uploadModel.deleteImgs(imagesProduct);
-      }
-      if (imageVariantsC && imageVariantsC.length > 0) {
-        await uploadModel.deleteImgs(imageVariantsC);
-      }
-      return res
-        .status(StatusCodes.BAD_REQUEST)
-        .json({ message: 'Có lỗi xảy ra xin thử lại sau' });
-    }
+    // return res.status(StatusCodes.OK).json({ parsedVariants });
+    const dataProduct = await productModel.update(id, data);
     if (dataProduct) {
       if (deleteImgs && deleteImgs.length > 0) {
-        await uploadModel.deleteImgs(deleteImgs);
+        uploadModel.deleteImgs(deleteImgs);
       }
 
       if (deleteImgsVars && deleteImgsVars.length > 0) {
-        await uploadModel.deleteImgs(deleteImgsVars);
+        uploadModel.deleteImgs(deleteImgsVars);
       }
 
       if (oldImageVariants && oldImageVariants.length > 0) {
-        await uploadModel.deleteImgs(oldImageVariants);
+        uploadModel.deleteImgs(oldImageVariants);
       }
-
-      if (thumbnail) {
-        await uploadModel.deleteImg(product.thumbnail);
-      }
-
       const result = dataProduct.result;
       if (!result) {
         return res
           .status(StatusCodes.NOT_FOUND)
           .json({ message: 'Không tìm thấy sản phẩm!' });
       }
-      return res.status(StatusCodes.OK).json({
-        result,
-      });
+      return res.status(StatusCodes.OK).json(result);
     }
   } catch (error) {
     if (req.files) {
-      const thumbnail = path.join(
-        'uploads/products',
-        req.files['thumbnail'][0].filename
-      );
+      const thumbnailPath = req.files['thumbnail']?.[0]?.filename;
 
-      let imagesProduct = req.files['images'].map((file) => {
+      const imageVariantsC = req.files['imageVariants']?.map((file) => {
         if (file && file.filename) {
           return path.join('uploads/products', file.filename);
         } else {
@@ -519,20 +456,36 @@ const updateProduct = async (req, res) => {
         }
       });
 
-      let imageVariantsC = req.files['imageVariants'].map((file) => {
+      const imagesProduct = req.files['images']?.map((file) => {
         if (file && file.filename) {
           return path.join('uploads/products', file.filename);
         } else {
           throw new Error('File image không hợp lệ');
         }
       });
-      uploadModel.deleteImg(thumbnail);
-      uploadModel.deleteImgs(imagesProduct);
-      uploadModel.deleteImgs(imageVariantsC);
+
+      if (thumbnailPath) {
+        const thumbnail = path.join('uploads/products', thumbnailPath);
+        uploadModel.deleteImg(thumbnail);
+      }
+
+      if (imagesProduct && imagesProduct.length > 0) {
+        uploadModel.deleteImgs(imagesProduct);
+      }
+
+      if (imageVariantsC && imageVariantsC.length > 0) {
+        uploadModel.deleteImgs(imageVariantsC);
+      }
     }
-    return res
-      .status(StatusCodes.INTERNAL_SERVER_ERROR)
-      .json({ message: error.message });
+    if (error.details) {
+      return res.status(StatusCodes.BAD_REQUEST).json({
+        message: error.details[0].message,
+      });
+    }
+    return res.status(StatusCodes.INTERNAL_SERVER_ERROR).json({
+      message: 'Có lỗi xảy ra xin thử lại sau',
+      error
+    });
   }
 };
